+++ lib/src/components/form/fields/date_picker.dart
@@ -11,6 +11,7 @@ import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/components/popover.dart';
 import 'package:shadcn_ui/src/raw_components/portal.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 import 'package:shadcn_ui/src/utils/gesture_detector.dart';
 import 'package:shadcn_ui/src/utils/states_controller.dart';
 
@@ -421,6 +422,12 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
 
     /// {@macro ShadButton.onFocusChange}
     ValueChanged<bool>? onFocusChange,
+
+    /// {@macro ShadButton.orderPolicy}
+    WidgetOrderPolicy? orderPolicy,
+
+    /// {@macro ShadButton.expands}
+    bool? expands,
   }) : super(
           builder: (field) {
             final state = field as _ShadFormBuilderDatePickerState;
@@ -557,6 +564,8 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
               textDirection: textDirection,
               onFocusChange: onFocusChange,
               placeholder: placeholder,
+              orderPolicy: orderPolicy,
+              expands: expands,
             );
           },
         );