+++ lib/src/components/radio.dart
@@ -7,6 +7,7 @@ import 'package:shadcn_ui/src/raw_components/focusable.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/debug_check.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 import 'package:shadcn_ui/src/utils/provider.dart';
 
 class ShadRadioGroup<T> extends StatefulWidget {
@@ -147,6 +148,7 @@ class ShadRadio<T> extends StatefulWidget {
     this.sublabel,
     this.padding,
     this.direction,
+    this.orderPolicy,
   });
 
   /// The value of the radio.
@@ -186,6 +188,12 @@ class ShadRadio<T> extends StatefulWidget {
   /// The direction of the radio.
   final TextDirection? direction;
 
+  /// {@template ShadRadio.orderPolicy}
+  /// The order policy of the items that compose the radio, defaults to
+  /// [WidgetOrderPolicy.linear()].
+  /// {@endtemplate}
+  final WidgetOrderPolicy? orderPolicy;
+
   @override
   State<ShadRadio<T>> createState() => _ShadRadioState<T>();
 }
@@ -211,7 +219,8 @@ class _ShadRadioState<T> extends State<ShadRadio<T>> {
   Widget build(BuildContext context) {
     assert(debugCheckHasShadTheme(context));
     final theme = ShadTheme.of(context);
-    final inheritedRadioGroup = context.watch<ShadRadioGroupState<dynamic>>();
+    final inheritedRadioGroup =
+        context.watch<ShadRadioGroupState<dynamic>>() as ShadRadioGroupState<T>;
 
     void onTap() {
       inheritedRadioGroup.select(widget.value);
@@ -237,6 +246,10 @@ class _ShadRadioState<T> extends State<ShadRadio<T>> {
     final effectivePadding = widget.padding ??
         theme.radioTheme.padding ??
         const EdgeInsets.only(left: 8);
+
+    final effectiveOrderPolicy = widget.orderPolicy ??
+        theme.radioTheme.orderPolicy ??
+        const WidgetOrderPolicy.linear();
 
     final radio = Semantics(
       checked: selected,
@@ -326,7 +339,7 @@ class _ShadRadioState<T> extends State<ShadRadio<T>> {
                   ),
                 ),
               ),
-          ],
+          ].order(effectiveOrderPolicy),
         ),
       ),
     );