+++ lib/src/theme/components/select.dart
@@ -4,6 +4,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:shadcn_ui/src/raw_components/portal.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 const kDefaultSelectMinWidth = 128.0;
 const kDefaultSelectMaxHeight = 384.0;
@@ -26,6 +27,7 @@ class ShadSelectTheme {
     this.effects,
     this.shadows,
     this.filter,
+    this.optionsOrderPolicy,
   });
 
   final bool merge;
@@ -50,6 +52,9 @@ class ShadSelectTheme {
   /// {@macro popover.shadows}
   final List<BoxShadow>? shadows;
 
+  /// {@macro ShadOption.orderPolicy}
+  final WidgetOrderPolicy? optionsOrderPolicy;
+
   static ShadSelectTheme lerp(
     ShadSelectTheme a,
     ShadSelectTheme b,
@@ -73,6 +78,7 @@ class ShadSelectTheme {
       clearSearchOnClose: t < 0.5 ? a.clearSearchOnClose : b.clearSearchOnClose,
       effects: t < 0.5 ? a.effects : b.effects,
       shadows: t < 0.5 ? a.shadows : b.shadows,
+      optionsOrderPolicy: t < .5 ? a.optionsOrderPolicy : b.optionsOrderPolicy,
     );
   }
 
@@ -94,6 +100,7 @@ class ShadSelectTheme {
     List<Effect<dynamic>>? effects,
     List<BoxShadow>? shadows,
     ImageFilter? filter,
+    WidgetOrderPolicy? optionsOrderPolicy,
   }) {
     return ShadSelectTheme(
       merge: merge ?? this.merge,
@@ -113,6 +120,7 @@ class ShadSelectTheme {
       effects: effects ?? this.effects,
       shadows: shadows ?? this.shadows,
       filter: filter ?? this.filter,
+      optionsOrderPolicy: optionsOrderPolicy ?? this.optionsOrderPolicy,
     );
   }
 
@@ -134,6 +142,7 @@ class ShadSelectTheme {
       effects: other.effects,
       shadows: other.shadows,
       filter: other.filter,
+      optionsOrderPolicy: other.optionsOrderPolicy,
     );
   }
 
@@ -156,7 +165,8 @@ class ShadSelectTheme {
         other.clearSearchOnClose == clearSearchOnClose &&
         other.effects == effects &&
         other.shadows == shadows &&
-        other.filter == filter;
+        other.filter == filter &&
+        other.optionsOrderPolicy == optionsOrderPolicy;
   }
 
   @override
@@ -175,6 +185,7 @@ class ShadSelectTheme {
         clearSearchOnClose.hashCode ^
         effects.hashCode ^
         shadows.hashCode ^
-        filter.hashCode;
+        filter.hashCode ^
+        optionsOrderPolicy.hashCode;
   }
 }