+++ lib/src/theme/components/switch.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/widgets.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 @immutable
 class ShadSwitchTheme {
@@ -16,6 +17,7 @@ class ShadSwitchTheme {
     this.duration,
     this.decoration,
     this.padding,
+    this.orderPolicy,
   });
 
   final bool merge;
@@ -38,6 +40,9 @@ class ShadSwitchTheme {
 
   final EdgeInsets? padding;
 
+  /// {@macro ShadSwitch.orderPolicy}
+  final WidgetOrderPolicy? orderPolicy;
+
   static ShadSwitchTheme lerp(
     ShadSwitchTheme a,
     ShadSwitchTheme b,
@@ -57,6 +62,7 @@ class ShadSwitchTheme {
       duration: b.duration,
       decoration: ShadDecoration.lerp(a.decoration, b.decoration, t),
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
+      orderPolicy: t < .5 ? a.orderPolicy : b.orderPolicy,
     );
   }
 
@@ -71,6 +77,7 @@ class ShadSwitchTheme {
     Duration? duration,
     ShadDecoration? decoration,
     EdgeInsets? padding,
+    WidgetOrderPolicy? orderPolicy,
   }) {
     return ShadSwitchTheme(
       merge: merge ?? this.merge,
@@ -83,6 +90,7 @@ class ShadSwitchTheme {
       duration: duration ?? this.duration,
       decoration: decoration ?? this.decoration,
       padding: padding ?? this.padding,
+      orderPolicy: orderPolicy ?? this.orderPolicy,
     );
   }
 
@@ -99,6 +107,7 @@ class ShadSwitchTheme {
       duration: other.duration,
       decoration: decoration?.mergeWith(other.decoration) ?? other.decoration,
       padding: other.padding,
+      orderPolicy: other.orderPolicy,
     );
   }
 
@@ -116,7 +125,8 @@ class ShadSwitchTheme {
         other.margin == margin &&
         other.duration == duration &&
         other.decoration == decoration &&
-        other.padding == padding;
+        other.padding == padding &&
+        other.orderPolicy == orderPolicy;
   }
 
   @override
@@ -130,6 +140,7 @@ class ShadSwitchTheme {
         margin.hashCode ^
         duration.hashCode ^
         decoration.hashCode ^
-        padding.hashCode;
+        padding.hashCode ^
+        orderPolicy.hashCode;
   }
 }